Think SMC is now working!
The notebook currently contains an example of fitting against a bivariate normal and fitting against the logistic model. Happy to split these into two notebooks if people think that'd be better but think we want both -- they illustrate how much you have to fiddle with the hyperparameters to get it to work!